refactor: add 93% ad a percentage and rename the file in pascal case
